JPERF-273: Hide untested new API
This is not a breaking change, because the new API wasn't released yet.
