--- src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/start/StartedJira.kt
@@ -1,8 +1,0 @@
-package com.atlassian.performance.tools.infrastructure.api.jira.start
-
-import com.atlassian.performance.tools.infrastructure.api.jira.install.InstalledJira
-
-class StartedJira(
-    val installed: InstalledJira,
-    val pid: Int
-)
