+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/jira/install/InstalledJira.kt
@@ -1,12 +1,10 @@
-package com.atlassian.performance.tools.infrastructure.api.jira.install
+package com.atlassian.performance.tools.infrastructure.jira.install
 
 import com.atlassian.performance.tools.infrastructure.api.jvm.JavaDevelopmentKit
 import com.atlassian.performance.tools.infrastructure.api.os.RemotePath
 
 /**
  * Points to an already installed Jira.
- *
- * @since 4.18.0
  */
 class InstalledJira(
     /**
