+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/jira/install/JiraInstallation.kt
@@ -1,10 +1,7 @@
-package com.atlassian.performance.tools.infrastructure.api.jira.install
+package com.atlassian.performance.tools.infrastructure.jira.install
 
 import net.jcip.annotations.ThreadSafe
 
-/**
- * @since 4.18.0
- */
 @ThreadSafe
 interface JiraInstallation {
