+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/jira/start/StartedJira.kt
@@ -1,0 +1,8 @@
+package com.atlassian.performance.tools.infrastructure.jira.start
+
+import com.atlassian.performance.tools.infrastructure.jira.install.InstalledJira
+
+class StartedJira(
+    val installed: InstalledJira,
+    val pid: Int
+)
